requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
rest_framework_simplejwt